#98 Adding Task Overview Chart To Dashboard
Added `Updated` property and Chart API endpoint

A new `Updated` property was added to the `Tasks` entity to track the last update time of a task. This property was also included in the `TaskContextModelSnapshot.cs` file and the `TaskController.cs` file was updated to set this property whenever a task is created or updated. The `Index.cshtml` file was updated to use the Chart.js library instead of ApexCharts for rendering charts. A new migration file was added to include the `Updated` column in the `Tasks` table. A new `ChartController.cs` file was added to provide an API endpoint that returns task data for charting. A `ChartDTO.cs` file was also added to define the data transfer object used by the new API endpoint.